requirements.txt
# Core dependencies
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
psycopg2-binary>=2.9.0

# Object Storage
boto3>=1.28.0

# Async support
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Logging
python-json-logger>=2.0.0

# Monitoring
prometheus-client>=0.17.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
